requirements.txt
# Core ML
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
d3rlpy>=2.0.0
econml>=0.14.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
google-cloud-bigquery>=3.10.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# API / Deployment
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
